refactor(api): split frontend client into namespace modules
Why: frontend/src/lib/api.ts had grown into a 1.8k-line monolith that was hard to navigate and merge safely.

What: extracted typed endpoint groups into frontend/src/lib/api/* modules, moved shared helpers into core.ts and types into types.ts, added api/index.ts barrel exports, and kept frontend/src/lib/api.ts as a compatibility re-export.

Tests: ran cd frontend; npm run lint (existing warnings only) and cd frontend; npm run build (passes).

Risk: Low to medium; import surface is preserved via barrel exports, but module resolution now depends on the new index wiring.
