+++ frontend/src/lib/api/index.ts
@@ -1,0 +1,22 @@
+export * from "./types";
+export { ApiError } from "./core";
+export { health } from "./health";
+export type { HealthResponse } from "./health";
+export { auth } from "./auth";
+export { student } from "./student";
+export { studentSubmissions } from "./studentSubmissions";
+export { notifications } from "./notifications";
+export { superadmin } from "./superadmin";
+export { orgs } from "./orgs";
+export { orgIntegrations } from "./orgIntegrations";
+export { userIntegrations } from "./userIntegrations";
+export { studentCourseGitHub } from "./studentCourseGitHub";
+export { staffCourseGitHub } from "./staffCourseGitHub";
+export { audit } from "./audit";
+export { orgUsers } from "./orgUsers";
+export { staff } from "./staff";
+export { users } from "./users";
+export { courseStaff } from "./courseStaff";
+export { staffSubmissions } from "./staffSubmissions";
+export { invites } from "./invites";
+export { playground } from "./playground";
